Refactored com.apgsga.microservice.patch.api.impl.PatchBean to com.apgsga.microservice.patch.api.Patch
