+++ .idea/jarRepositories.xml
@@ -21,5 +21,10 @@
       <option name="name" value="maven" />
       <option name="url" value="https://artifactory4t4apgsga.jfrog.io/artifactory4t4apgsga/repo" />
     </remote-repository>
+    <remote-repository>
+      <option name="id" value="maven" />
+      <option name="name" value="maven" />
+      <option name="url" value="https://artifactory4t4apgsga.jfrog.io/artifactory/repo" />
+    </remote-repository>
   </component>
 </project>
