+++ apg-patch-service-cmdclient/build.gradle
@@ -2,7 +2,7 @@ import org.apache.tools.ant.filters.*;
 plugins {
   	id 'groovy'
     id 'application'
-	id 'com.github.johnrengelman.shadow' version '2.0.2'
+	id 'com.github.johnrengelman.shadow' version '5.2.0'
   	id 'nebula.ospackage' version '4.10.0'
   	id 'ivy-publish' 
 }
@@ -185,5 +185,6 @@ artifactory {
 
 assemble { dependsOn copyLib,pkgAsGTar}
 buildRpm {dependsOn processResources, assemble}
+publish {dependsOn buildRpm}
