+++ apg-patch-service-cmdclient/src/test/groovy/com/apgsga/patch/service/client/DbCliIntegrationTest.groovy
@@ -61,8 +61,9 @@ class DbCliIntegrationTest extends Specification {
 		cleanup:
 		outputFile.delete()
 	}
-	
-	
+
+	// TODO (jhe, 8.4) The user does'nt exist
+	@Ignore
 	@Requires({patchExists("5799")})
 	def "Patch DB Cli  update status of Patch"() {
 		when:
