+++ apg-patch-service-cmdclient/src/test/resources/Patch5791.json
@@ -1,1 +1,1 @@
-{"className":"com.apgsga.microservice.patch.api.impl.PatchBean","patchNummer":"5791","serviceName":"It21Ui","microServiceBranch":"it21_release_9_1_0_admin_uimig","dbPatchBranch":"Patch_0900C1_5791","prodBranch":"HEAD","patchTag":"Patch_0900C1_5791_1","tagNr":1,"installationTarget":"CHEI212","baseVersionNumber":"9.1.0","revisionMnemoPart":"ADMIN-UIMIG","revisionNumber":null,"lastRevisionNumber":null,"dbObjects":[],"mavenArtifacts":[{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"alog-ui","groupId":"com.affichage.it21.alog","name":"com.affichage.it21.alog.ui","version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0}],"installOnEmptyModules":false,"mavenArtifactsAsVcsPath":["com.affichage.it21.alog.ui"],"installJadasAndGui":true,"mavenArtifactsToBuild":[{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"alog-ui","groupId":"com.affichage.it21.alog","name":"com.affichage.it21.alog.ui","version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0}],"dbObjectsAsVcsPath":[]}
+{"className":"com.apgsga.microservice.patch.api.Patch","patchNummer":"5791","serviceName":"It21Ui","microServiceBranch":"it21_release_9_1_0_admin_uimig","dbPatchBranch":"Patch_0900C1_5791","prodBranch":"HEAD","patchTag":"Patch_0900C1_5791_1","tagNr":1,"installationTarget":"CHEI212","baseVersionNumber":"9.1.0","revisionMnemoPart":"ADMIN-UIMIG","revisionNumber":null,"lastRevisionNumber":null,"dbObjects":[],"mavenArtifacts":[{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"alog-ui","groupId":"com.affichage.it21.alog","name":"com.affichage.it21.alog.ui","version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0}],"installOnEmptyModules":false,"mavenArtifactsAsVcsPath":["com.affichage.it21.alog.ui"],"installJadasAndGui":true,"mavenArtifactsToBuild":[{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"alog-ui","groupId":"com.affichage.it21.alog","name":"com.affichage.it21.alog.ui","version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0}],"dbObjectsAsVcsPath":[]}
