+++ apg-patch-service-common/src/main/java/com/apgsga/microservice/patch/api/Service.java
@@ -1,0 +1,9 @@
+package com.apgsga.microservice.patch.api;
+
+import com.affichage.persistence.common.client.EntityRootInterface;
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
+
+@EntityRootInterface
+@JsonTypeInfo(use = JsonTypeInfo.Id.CLASS, property = "className")
+public class Service {
+}
