+++ apg-patch-service-server/db/Patch5401.json
@@ -1,1 +1,1 @@
-{"className":"com.apgsga.microservice.patch.api.impl.PatchBean","patchNummer":"5401","serviceName":"It21Ui","microServiceBranch":"it21_release_9_1_0_admin_uimig","dbPatchBranch":"Patch_0900C1_5401","prodBranch":"HEAD","patchTag":"Patch_0900C1_5401_1","developerBranch":"","tagNr":1,"installationTarget":"CHEI212","targetToState":null,"baseVersionNumber":null,"revisionMnemoPart":null,"revision":null,"lastRevision":null,"runningNr":null,"dbObjects":[{"className":"com.apgsga.microservice.patch.api.impl.DbObjectBean","fileName":"apgops_lib.apg_db_outage_planned.syn.sql","filePath":"testdbmodule/database/sql/","moduleName":"testdbmodule"},{"className":"com.apgsga.microservice.patch.api.impl.DbObjectBean","fileName":"sys.apg_install_utility.pks.cleanup.sql","filePath":"testdbmodule/database/sql/","moduleName":"testdbmodule"},{"className":"com.apgsga.microservice.patch.api.impl.DbObjectBean","fileName":"test.sys.apg_install_utility.pks.cleanup.sql","filePath":"testdbAnotherdbModule/database/sql/","moduleName":"testdbAnotherdbModule"}],"mavenArtifacts":[{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"zentraldispo-dao","groupId":"com.affichage.it21.vk","name":null,"version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0}],"installOnEmptyModules":false,"lastPipelineTask":"","currentTarget":null,"currentPipelineTask":null,"logText":null,"dbObjectsAsVcsPath":["testdbmodule/database/sql//apgops_lib.apg_db_outage_planned.syn.sql","testdbmodule/database/sql//sys.apg_install_utility.pks.cleanup.sql","testdbAnotherdbModule/database/sql//test.sys.apg_install_utility.pks.cleanup.sql"],"mavenArtifactsAsVcsPath":[null],"mavenArtifactsToBuild":[{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"zentraldispo-dao","groupId":"com.affichage.it21.vk","name":null,"version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0}],"installJadasAndGui":true}
+{"className":"com.apgsga.microservice.patch.api.Patch","patchNummer":"5401","serviceName":"It21Ui","microServiceBranch":"it21_release_9_1_0_admin_uimig","dbPatchBranch":"Patch_0900C1_5401","prodBranch":"HEAD","patchTag":"Patch_0900C1_5401_1","developerBranch":"","tagNr":1,"installationTarget":"CHEI212","targetToState":null,"baseVersionNumber":null,"revisionMnemoPart":null,"revision":null,"lastRevision":null,"runningNr":null,"dbObjects":[{"className":"com.apgsga.microservice.patch.api.impl.DbObjectBean","fileName":"apgops_lib.apg_db_outage_planned.syn.sql","filePath":"testdbmodule/database/sql/","moduleName":"testdbmodule"},{"className":"com.apgsga.microservice.patch.api.impl.DbObjectBean","fileName":"sys.apg_install_utility.pks.cleanup.sql","filePath":"testdbmodule/database/sql/","moduleName":"testdbmodule"},{"className":"com.apgsga.microservice.patch.api.impl.DbObjectBean","fileName":"test.sys.apg_install_utility.pks.cleanup.sql","filePath":"testdbAnotherdbModule/database/sql/","moduleName":"testdbAnotherdbModule"}],"mavenArtifacts":[{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"zentraldispo-dao","groupId":"com.affichage.it21.vk","name":null,"version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0}],"dockerServices":[],"installDockerServices":false,"installOnEmptyModules":false,"lastPipelineTask":"","currentTarget":null,"currentPipelineTask":null,"logText":null,"dbObjectsAsVcsPath":["testdbmodule/database/sql//apgops_lib.apg_db_outage_planned.syn.sql","testdbmodule/database/sql//sys.apg_install_utility.pks.cleanup.sql","testdbAnotherdbModule/database/sql//test.sys.apg_install_utility.pks.cleanup.sql"],"mavenArtifactsAsVcsPath":[null],"mavenArtifactsToBuild":[{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"zentraldispo-dao","groupId":"com.affichage.it21.vk","name":null,"version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0}],"installJadasAndGui":true}
