+++ apg-patch-service-server/db/Patch5402.json
@@ -1,1 +1,1 @@
-{"className":"com.apgsga.microservice.patch.api.impl.PatchBean","patchNummer":"5402","serviceName":"It21Ui","microServiceBranch":"it21_release_9_1_0_admin_uimig","dbPatchBranch":"Patch_0900C1_5402","prodBranch":"HEAD","patchTag":"Patch_0900C1_5402_2","developerBranch":"","tagNr":2,"installationTarget":"CHEI212","targetToState":null,"baseVersionNumber":"XXXX","revisionMnemoPart":"ADMIN-UIMIG","revision":null,"lastRevision":null,"runningNr":null,"dbObjects":[{"className":"com.apgsga.microservice.patch.api.impl.DbObjectBean","fileName":"FileName1","filePath":"FilePath1","moduleName":null}],"mavenArtifacts":[{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"zentraldispo-dao","groupId":"com.affichage.it21.vk","name":"com.affichage.it21.vk.zentraldispo.dao","version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0},{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"zentraldispo-ui","groupId":"com.affichage.it21.vk","name":"com.affichage.it21.vk.zentraldispo.ui","version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0}],"installOnEmptyModules":false,"lastPipelineTask":"","currentTarget":null,"currentPipelineTask":null,"logText":null,"dbObjectsAsVcsPath":["FilePath1/FileName1"],"mavenArtifactsAsVcsPath":["com.affichage.it21.vk.zentraldispo.dao","com.affichage.it21.vk.zentraldispo.ui"],"mavenArtifactsToBuild":[{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"zentraldispo-dao","groupId":"com.affichage.it21.vk","name":"com.affichage.it21.vk.zentraldispo.dao","version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0},{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"zentraldispo-ui","groupId":"com.affichage.it21.vk","name":"com.affichage.it21.vk.zentraldispo.ui","version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0}],"installJadasAndGui":true}
+{"className":"com.apgsga.microservice.patch.api.Patch","patchNummer":"5402","serviceName":"It21Ui","microServiceBranch":"it21_release_9_1_0_admin_uimig","dbPatchBranch":"Patch_0900C1_5402","prodBranch":"HEAD","patchTag":"Patch_0900C1_5402_2","developerBranch":"","tagNr":2,"installationTarget":"CHEI212","targetToState":null,"baseVersionNumber":"XXXX","revisionMnemoPart":"ADMIN-UIMIG","revision":null,"lastRevision":null,"runningNr":null,"dbObjects":[{"className":"com.apgsga.microservice.patch.api.impl.DbObjectBean","fileName":"FileName1","filePath":"FilePath1","moduleName":null}],"mavenArtifacts":[{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"zentraldispo-dao","groupId":"com.affichage.it21.vk","name":"com.affichage.it21.vk.zentraldispo.dao","version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0},{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"zentraldispo-ui","groupId":"com.affichage.it21.vk","name":"com.affichage.it21.vk.zentraldispo.ui","version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0}],"dockerServices":[],"installDockerServices":false,"installOnEmptyModules":false,"lastPipelineTask":"","currentTarget":null,"currentPipelineTask":null,"logText":null,"dbObjectsAsVcsPath":["FilePath1/FileName1"],"mavenArtifactsAsVcsPath":["com.affichage.it21.vk.zentraldispo.dao","com.affichage.it21.vk.zentraldispo.ui"],"mavenArtifactsToBuild":[{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"zentraldispo-dao","groupId":"com.affichage.it21.vk","name":"com.affichage.it21.vk.zentraldispo.dao","version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0},{"className":"com.apgsga.microservice.patch.api.impl.MavenArtifactBean","artifactId":"zentraldispo-ui","groupId":"com.affichage.it21.vk","name":"com.affichage.it21.vk.zentraldispo.ui","version":"9.1.0.ADMIN-UIMIG-SNAPSHOT","dependencyLevel":0}],"installJadasAndGui":true}
