+++ apg-patch-service-server/src/test/java/com/apgsga/microservice/patch/server/MicroServicePatchServerExceptionTests.java
@@ -20,17 +20,13 @@ import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import com.apgsga.microservice.patch.api.Patch;
-import com.apgsga.microservice.patch.api.PatchLog;
 import com.apgsga.microservice.patch.api.PatchPersistence;
 import com.apgsga.microservice.patch.api.impl.MavenArtifactBean;
-import com.apgsga.microservice.patch.api.impl.PatchBean;
-import com.apgsga.microservice.patch.api.impl.PatchLogBean;
 import com.apgsga.microservice.patch.exceptions.PatchServiceRuntimeException;
 import com.apgsga.microservice.patch.server.impl.GroovyScriptActionExecutor;
 import com.apgsga.microservice.patch.server.impl.PatchActionExecutor;
 import com.apgsga.microservice.patch.server.impl.PatchActionExecutorFactory;
 import com.apgsga.microservice.patch.server.impl.SimplePatchContainerBean;
-import com.jcraft.jsch.Logger;
 
 @RunWith(SpringRunner.class)
 @DirtiesContext(classMode = ClassMode.BEFORE_CLASS)
@@ -52,7 +48,7 @@ public class MicroServicePatchServerExceptionTests {
 	private PatchActionExecutorFactory patchActionFactory;
 
 	@Before
-	public void setUp() throws IOException {
+	public void setUp() {
 		repo.clean();
 	}
 
@@ -104,7 +100,7 @@ public class MicroServicePatchServerExceptionTests {
 	
 	@Test
 	public void testSaveEmptyWithOutId() {
-		Patch patch = new PatchBean();
+		Patch patch = new Patch();
 		try {
 			patchService.save(patch);
 			fail();
@@ -127,7 +123,7 @@ public class MicroServicePatchServerExceptionTests {
 	
 	@Test
 	public void testRemoveEmptyWithOutId() {
-		Patch patch = new PatchBean();
+		Patch patch = new Patch();
 		try {
 			patchService.remove(patch);
 			fail();
@@ -139,7 +135,7 @@ public class MicroServicePatchServerExceptionTests {
 	
 	@Test
 	public void testRemoveDoesnotExist() {
-		Patch patch = new PatchBean();
+		Patch patch = new Patch();
 		patch.setPatchNummer("XXXX");
 		try {
 			patchService.remove(patch);
@@ -163,7 +159,7 @@ public class MicroServicePatchServerExceptionTests {
 	
 	@Test
 	public void testPatchInvalidToState() {
-		Patch patch = new PatchBean();
+		Patch patch = new Patch();
 		patch.setPatchNummer("SomeUnqiueNumber3");
 		patch.setServiceName("It21ui");
 		patch.setMicroServiceBranch("SomeBaseBranch");
@@ -182,7 +178,7 @@ public class MicroServicePatchServerExceptionTests {
 	
 	@Test
 	public void testPatchConfigDirDoesnotExist() {
-		Patch patch = new PatchBean();
+		Patch patch = new Patch();
 		patch.setPatchNummer("SomeUnqiueNumber3");
 		patch.setServiceName("It21ui");
 		patch.setMicroServiceBranch("SomeBaseBranch");
@@ -202,7 +198,7 @@ public class MicroServicePatchServerExceptionTests {
 	
 	@Test
 	public void testPatchConfigFileDoesnotExist() {
-		Patch patch = new PatchBean();
+		Patch patch = new Patch();
 		patch.setPatchNummer("SomeUnqiueNumber3");
 		patch.setServiceName("It21ui");
 		patch.setMicroServiceBranch("SomeBaseBranch");
@@ -223,7 +219,7 @@ public class MicroServicePatchServerExceptionTests {
 	
 	@Test
 	public void testPatchInvalidGroovyScriptFile() {
-		Patch patch = new PatchBean();
+		Patch patch = new Patch();
 		patch.setPatchNummer("SomeUnqiueNumber3");
 		patch.setServiceName("It21ui");
 		patch.setMicroServiceBranch("SomeBaseBranch");
